+++ src/app/core/components/forms/new-player-form/new-player-form.component.html
