--- src/app/core/components/new-match-form/new-match-form.component.html
@@ -1,49 +1,0 @@
-<card-wrapper>
-    <h1 class="text-center">Nuevo partido</h1>
-
-    <form [formGroup]="newMatchForm" (ngSubmit)="onSubmit()">
-        <div class="mb-3">
-            <label [for]="controls.dateTime" class="form-label">Fecha y Hora</label>
-            <input
-                type="dateTime-local"
-                [min]="getDateNow()"
-                class="form-control {{ getControlValidClass(controls.dateTime) }}"
-                [id]="controls.dateTime"
-                [formControl]="controls.dateTime"
-            />
-            <div *ngIf="controlHasError(controls.dateTime)" class="invalid-feedback">
-                La fecha y hora es obligatoria
-            </div>
-        </div>
-
-        <div class="mb-3">
-            <label for="location" class="form-label">Lugar</label>
-            <input
-                placeholder="Ubicación del partido"
-                type="text"
-                class="form-control {{ getControlValidClass(controls.location) }}"
-                id="location"
-                [formControl]="controls.location"
-            />
-            <div *ngIf="controlHasError(controls.location, 'required')" class="invalid-feedback">
-                El lugar es obligatorio
-            </div>
-            <div *ngIf="controlHasError(controls.location, 'minlength')" class="invalid-feedback">
-                El lugar debe tener al menos 4 caracteres
-            </div>
-        </div>
-
-        <div class="d-flex justify-content-between">
-            <button type="button" class="btn btn-outline-secondary" (click)="goHome()">
-                Volver a inicio
-            </button>
-            <button type="submit" class="btn btn-primary">Crear partido</button>
-        </div>
-    </form>
-
-    <div class="alert alert-success mt-4" *ngIf="resultMessage">
-        <p class="text-center">
-            {{ resultMessage }}
-        </p>
-    </div>
-</card-wrapper>
