+++ src/app/core/core-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './components/home/home.component';
-import { NewMatchFormComponent } from './components/new-match-form/new-match-form.component';
-import { NewPlayerFormComponent } from './components/new-player-form/new-player-form.component';
-import { MatchDetailsFormComponent } from './components/match-details-form/match-details-form.component';
-import { StatisticsFormComponent } from './components/statistics-form/statistics-form.component';
+import { NewMatchFormComponent } from './components/forms/new-match-form/new-match-form.component';
+import { NewPlayerFormComponent } from './components/forms/new-player-form/new-player-form.component';
+import { MatchDetailsFormComponent } from './components/forms/match-details-form/match-details-form.component';
+import { StatisticsFormComponent } from './components/forms/statistics-form/statistics-form.component';
 
 const routes: Routes = [
   {
