+++ src/app/core/core.module.ts
@@ -6,12 +6,12 @@ import { HomeComponent } from './components/home/home.component';
 import { HttpClientModule } from '@angular/common/http';
 import { LayoutModule } from '@angular/cdk/layout';
 import { ReactiveFormsModule } from '@angular/forms';
-import { NewPlayerFormComponent } from './components/new-player-form/new-player-form.component';
-import { NewMatchFormComponent } from './components/new-match-form/new-match-form.component';
-import { MatchDetailsFormComponent } from './components/match-details-form/match-details-form.component';
+import { NewPlayerFormComponent } from './components/forms/new-player-form/new-player-form.component';
+import { NewMatchFormComponent } from './components/forms/new-match-form/new-match-form.component';
+import { MatchDetailsFormComponent } from './components/forms/match-details-form/match-details-form.component';
 import { CardWrapperComponent } from './components/card-wrapper/card-wrapper.component';
 import { FormInputComponent } from './components/form-input/form-input.component';
-import { StatisticsFormComponent } from './components/statistics-form/statistics-form.component';
+import { StatisticsFormComponent } from './components/forms/statistics-form/statistics-form.component';
 
 @NgModule({
   declarations: [
